LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/sandboxed-iframe-fetch-event.https-expected.txt

CONSOLE MESSAGE: line 27: [blocked] The page at https://localhost:9443/service-workers/service-worker/sandboxed-iframe-fetch-event.https.html was not allowed to display insecure content from blob:null/2548f006-c6d1-4d09-82c1-3c1b434dbfe7.



Harness Error (TIMEOUT), message = null

TIMEOUT ServiceWorker FetchEvent for sandboxed iframe. Test timed out
